examples/start-agents/dspy_starter/requirements.txt
dspy>=2.5.0
jupyter>=1.0.0
python-dotenv>=1.1.0
streamlit>=1.32.0